fix: input and output box should stick on top
when the output or another div extends the parent div's height, the other div (this case it's input) would be centered.

one way to achieve this is to prompt with a simple code and translate it into assembly (any language that would produce more lines of code than the input), then hope for the output to exceed the input div's height.